fix-install: Remove 'sudo' commands in fix-install
This script should only be run as root, and should fail in all other cases.

In a stock Debian environment, the 'sudo' command does not exist, and the fix-install script will fail when attempting to call 'sudo'. Installation of 'sudo' will fix the problem, however if we are not intending to install 'sudo' and use a stock Debian type of setup, then this is a wrong approach. This also applies to other environments which may not even have the 'sudo' command available to it.